+++ docs/src/pages/index.astro
@@ -35,7 +35,7 @@ import "../styles/base.css";
 		<ThemeToggleButton client:load />
 	</header>
 	<section class="w-full max-w-7xl mx-auto pt-6 pb-10 px-4">
-		<h1 class="mb-4 text-center font-semibold text-4xl">Bring Radix Colors to Tailwind CSS</h1>
+		<h1 class="mb-4 text-center font-semibold text-mauve-12 text-4xl">Bring Radix Colors to Tailwind CSS</h1>
 		<p class="mb-8 text-center text-lg text-mauve-11">Improving the compatibility of Radix Colors and Tailwind CSS.</p>
 		<div class="flex items-center justify-center">
 			<code class="rounded-full bg-slate-3 text-sm text-mauve-12 px-3 py-1">
